+++ etna/datasets/tsdataset.py
@@ -1296,7 +1296,7 @@ class TSDataset:
         Raises
         ------
         ValueError:
-            If ``features`` list contains target components
+            If ``features`` list contains target or target components
         """
         features_set = set(features)
 
@@ -1311,6 +1311,9 @@ class TSDataset:
             raise ValueError(
                 "Prediction intervals can't be dropped from the dataset using this method! Use `drop_prediction_intervals` method!"
             )
+
+        if "target" in features_set:
+            raise ValueError(f"Target can't be dropped from the dataset!")
 
         dfs = [("df", self.df)]
         if drop_from_exog: